# energy-consumption-monitoring/docs/api-spec.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Energy Consumption Monitoring API",
        "version": "1.0.0",
        "description": "校园智慧后勤能耗监测系统后端 API 文档。"
    },
    "paths": {
...
        "/api/devices/{id}/bind-room/": {
            "post": {
                "operationId": "api_devices_bind_room_create",
                "description": "为设备绑定或解绑房间。",
                "summary": "绑定设备房间",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "标识此 device 的 唯一整数值。",
                        "required": true
                    }
                ],
                "tags": [
                    "api"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BindRoomRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/BindRoomRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/BindRoomRequest"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeviceDetail"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "房间不存在"
                    }
                }
            }
        },
        "/api/devices/data-status/": {
            "get": {
                "operationId": "api_devices_data_status_retrieve",
                "description": "返回设备是否有采集数据及最近数据时间。",
                "summary": "获取设备数据状态",
                "tags": [
                    "api"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {}
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BindRoomRequest": {
                "type": "object",
                "properties": {
                    "room_id": {
                        "type": "integer",
                        "nullable": true
                    }
                }
            },
...
            "CodeEnum": {
                "enum": [
                    "WATER",
                    "ELECTRICITY",
                    "GAS"
                ],
                "type": "string",
                "description": "* `WATER` - Water\n* `ELECTRICITY` - Electricity\n* `GAS` - Gas"
            },
...
            "DeviceDetail": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "device_id": {
                        "type": "string",
                        "maxLength": 64
                    },
                    "name": {
                        "type": "string",
                        "title": "Device name",
                        "maxLength": 128
                    },
                    "energy_type": {
                        "type": "integer"
                    },
                    "energy_type_detail": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/EnergyType"
                            }
                        ],
                        "readOnly": true
                    },
                    "room": {
                        "type": "integer",
                        "nullable": true
                    },
                    "room_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "floor_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "building_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "campus_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "model": {
                        "type": "string",
                        "nullable": true,
                        "title": "Device model",
                        "maxLength": 64
                    },
                    "status": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/StatusC1dEnum"
                            }
                        ],
                        "title": "Device status"
                    },
                    "last_data_time": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "latest_data": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "building_name",
                    "campus_name",
                    "created_at",
                    "device_id",
                    "energy_type",
                    "energy_type_detail",
                    "floor_name",
                    "id",
                    "latest_data",
                    "name",
                    "room_name",
                    "updated_at"
                ]
            },
...
            "EnergyType": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string",
                        "title": "Energy name",
                        "maxLength": 64
                    },
                    "code": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/CodeEnum"
                            }
                        ],
                        "title": "Energy code"
                    },
                    "unit": {
                        "type": "string",
                        "maxLength": 32
                    },
                    "icon": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 128
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "code",
                    "created_at",
                    "id",
                    "name",
                    "unit",
                    "updated_at"
                ]
            },
...
            "StatusC1dEnum": {
                "enum": [
                    "ONLINE",
                    "OFFLINE",
                    "FAULT"
                ],
                "type": "string",
                "description": "* `ONLINE` - Online\n* `OFFLINE` - Offline\n* `FAULT` - Fault"
            },
...
        },
        "securitySchemes": {
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}